+++ src/constants/logger.enum.ts
@@ -1,0 +1,20 @@
+/**
+ * Copyright (C) 2020 - 2021 IKOA Business Opportunity
+ *
+ * All Rights Reserved
+ * Author: Reinier Millo Sánchez <millo@ikoabo.com>
+ *
+ * This file is part of the IKOA Business Oportunity Server Package
+ * It can't be copied and/or distributed without the express
+ * permission of the author.
+ */
+
+export const LOG_LEVEL = {
+  ERROR: "error",
+  WARN: "warn",
+  INFO: "info",
+  HTTP: "http",
+  VERBOSE: "verbose",
+  DEBUG: "debug",
+  SILLY: "silly"
+};